requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.1
django-ckeditor==6.7.2
django-js-asset==3.1.2
pillow==11.2.1
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
